One additional intptr_t conversion (in the same vein as check-in 4824) (CVS 4859)
FossilOrigin-Name: 2777ca788204c37ae5607f7acabe78f414255fa2

+++ src/vdbemem.c
@@ -932,7 +932,7 @@ const void *sqlite3ValueText(sqlite3_value* pVal, u8 enc){
   }else{
     assert( (pVal->flags&MEM_Blob)==0 );
     sqlite3VdbeMemStringify(pVal, enc);
-    assert( 0==(1&(int)pVal->z) );
+    assert( 0==(1&(intptr_t)pVal->z) );
   }
   assert(pVal->enc==(enc & ~SQLITE_UTF16_ALIGNED) || pVal->db==0
               || pVal->db->mallocFailed );
